+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- An error message is displayed when the report task to redmine fails
+
 ## [0.1.2] - 2023-06-03
 
 ### Added