+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Show task preview after modify it
 - Allow set db and integration through config file
 - Help improved
+- Add some unit tests
 
 ## [0.1.2] - 2023-06-03